+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-## [4.17.5] - 2026-09-07
+## [4.17.5] - 2026-09-15
 
 * [PR-683](https://github.com/itk-dev/deltag.aarhus.dk/pull/683)
   * Fixed dialogue bug in timeline view